Refactor: utils에서 lib으로 이동
